refactor: BottomFixedContext 파일 분리
