+++ frontend/src/shared/ui/bottom-fixed/BottomFixedProvider.tsx
@@ -1,0 +1,10 @@
+import { ProviderProps } from 'react';
+import { ContextValue } from './type';
+import { Context } from './BottomFixedContext';
+
+export default function BottomFixedProvider({
+  value,
+  children,
+}: ProviderProps<ContextValue>) {
+  return <Context.Provider value={value}>{children}</Context.Provider>;
+}
