+++ frontend/src/shared/ui/bottom-fixed/type.ts
@@ -1,0 +1,3 @@
+export type ContextValue = {
+  bottomPosition: number;
+};
